requirements.txt
accelerate==0.23.0
asyncio==3.4.3
bitsandbytes==0.41.1
certifi
fastapi==0.109.2
fastapi-utils==0.2.1
huggingface-hub==0.20.3
protobuf==4.24.4
python-dotenv==1.0.0
requests==2.31.0
sentencepiece==0.1.99
starlette==0.36.3
scipy==1.11.3
torch==2.1.1
transformers==4.36.0
urllib3==2.0.7
uvicorn==0.19.0
wheel==0.38.1